implement essencial subgraph queries
