+++ packages/web-app/src/queries/dao.tsx
@@ -1,44 +1,9 @@
 import {gql} from '@apollo/client';
 
-export const DAO_ERC20VOTING_PACKAGE_BY_ID = gql`
-  query erc20VotingPackages($id: ID) {
-    erc20VotingPackages(where: {id: $id}) {
-      id
-      proposals {
-        id
-      }
-      supportRequiredPct
-      participationRequiredPct
-      minDuration
-      votesLength
-      token {
-        id
-        name
-        symbol
-        decimals
-      }
-    }
-  }
-`;
-
-export const DAO_WHITELIST_PACKAGE_BY_ID = gql`
-  query whitelistPackages($id: ID) {
-    whitelistPackages(where: {id: $id}) {
-      id
-      proposals {
-        id
-      }
-      supportRequiredPct
-      participationRequiredPct
-      minDuration
-      votesLength
-      users {
-        id
-      }
-    }
-  }
-`;
-
+/**
+ * from actions to proposals can be paginanted as needed
+ * and or can be used to get array length
+ */
 export const DAO_LIST = gql`
   query DAOs($offset: Int, $limit: Int) {
     daos(
@@ -85,6 +50,9 @@ export const DAO_LIST = gql`
   }
 `;
 
+/**
+ * from actions to proposals can be paginanted as needed
+ */
 export const DAO_BY_NAME = gql`
   query DAO($name: String) {
     daos(where: {name: $name}) {
